planetgen_engine: Don't create GL buffers per mesh
Instead we create a small number of large buffers which we then suballocate
from. This allows us to do fewer buffer/VAO binds giving a nice performance
boost; it should also be more similar to how we would do things when using
Vulkan.

There are still further improvements that could be made, but this is a solid
basis to work from.