planetgen: Implement "horizon culling"
Implements what I call "horizon culling". Horizon culling determines which quads
don't need to be rendered based on if they guaranteed to be over the horizon. It
requires the quad-sphere to have a minimum and maximum height, and the closer
these heights are the better the results (e.g. we can't assume much if the
"planet" could be literally pear shaped).