planetgen_engine: Experimental buffer update method
Experimental method that tries to avoid synchronization, the idea is that we
abandon old allocations when we need to update a buffer and use an unallocated
region. Fences are used to ensure that we are truly finished with a range before
we actually return it to the allocation system. However for a reason that I do
not understand I had rendering artefacts initially; this was (hopefully?) solved
by moving the fence creation to after buffer updates (instead of before). I
don't understand why this works now but not before.

This commit can be reverted (wholly or partially) if need be.